requirements.txt
fastapi==0.116.1
pyjwt==2.10.1
uvicorn==0.35.0
python-multipart==0.0.20
passlib
